refactor: seperate the layout for each section
