+++ client/app/(dashboard)/client/[id]/page.jsx
