+++ client/app/payment/[id]/page.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Page = () => {
+  return <div>Page</div>;
+};
+
+export default Page;
