+++ client/utils/fonts.js
@@ -1,0 +1,7 @@
+import { Poppins } from "next/font/google";
+
+export const poppins = Poppins({
+  subsets: ["latin"],
+  weight: ["600", "700", "400"],
+  variable: "--font-poppins",
+});
